requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-text-splitters
langchain-groq
langchain-huggingface
huggingface_hub
chromadb
pypdf
python-dotenv
requests
python-multipart
streamlit
pymupdf